[HMTM-202] - GQL-Client architecture update
