+++ graphql_client/__init__.py
@@ -1,66 +1,3 @@
-from .sso import (
-    RegisterUserVariables,
-    RegisterUserMutation,
-
-    LoginUserVariables,
-    LoginUserMutation,
-
-    VerifyUserEmailVariables,
-    VerifyUserEmailMutation,
-
-    GetMeQuery,
-
-    ToysCounterQuery,
-
-    AllToysCategoriesQuery,
-    AllToysTagsQuery,
-
-    GetMasterByUserQuery,
-    GetMasterByUserVariables,
-
-    GetUserByIDQuery,
-    GetUserByIDVariables,
-
-    GetUserByEmailQuery,
-    GetUserByEmailVariables,
-
-    ToysCatalogQuery,
-    ToysCatalogVariables,
-    ToysCounterFiltersVariables,
-
-    ToyByIDQuery,
-    ToyByIDVariables,
-
-    RefreshTokensMutation,
-
-    SendVerifyEmailMessageVariables,
-    SendVerifyEmailMessageMutation,
-
-    SendForgetPasswordMessageVariables,
-    SendForgetPasswordMessageMutation,
-
-    ChangePasswordVariables,
-    ChangePasswordMutation,
-
-    ForgetPasswordVariables,
-    ChangeForgetPasswordMutation,
-
-    UpdateUserProfileVariables,
-    UpdateUserProfileMutation,
-
-    UpdateMasterVariables,
-    UpdateMasterMutation,
-
-    RegisterMasterVariables,
-    RegisterMasterMutation,
-
-    MastersCatalogVariables,
-    MastersCatalogQuery,
-
-    MastersCounterQuery,
-    MasterCounterVariables
-)
-
 from .client import GraphQLClient
 from .errors_handlers import extract_error_message
 from .response_processor import ResponseProcessor
