+++ graphql_client/queries/common.py
@@ -1,0 +1,24 @@
+from gql import gql
+from graphql import DocumentNode
+
+
+class GetMeQuery:
+    @staticmethod
+    def to_gql() -> DocumentNode:
+        return gql("""
+            query {
+                me {
+                    id
+                    displayName
+                    email
+                    emailConfirmed
+                    phone
+                    phoneConfirmed
+                    telegram
+                    telegramConfirmed
+                    avatar
+                    createdAt
+                    updatedAt
+                    }
+                }
+            """)
