--- graphql_client/sso/__init__.py
@@ -1,46 +1,0 @@
-from .mutations import (
-    RegisterUserMutation,
-    LoginUserMutation,
-    VerifyUserEmailMutation,
-    RefreshTokensMutation,
-    SendVerifyEmailMessageMutation,
-    SendForgetPasswordMessageMutation,
-    ChangePasswordMutation,
-    ChangeForgetPasswordMutation,
-    UpdateUserProfileMutation,
-    UpdateMasterMutation,
-    RegisterMasterMutation,
-)
-from .variables import (
-    RegisterUserVariables,
-    LoginUserVariables,
-    VerifyUserEmailVariables,
-    SendVerifyEmailMessageVariables,
-    SendForgetPasswordMessageVariables,
-    ChangePasswordVariables,
-    ForgetPasswordVariables,
-    UpdateUserProfileVariables,
-    UpdateMasterVariables,
-    RegisterMasterVariables,
-    GetMasterByUserVariables,
-    GetUserByIDVariables,
-    GetUserByEmailVariables,
-    ToysCatalogVariables,
-    ToyByIDVariables,
-    MastersCatalogVariables,
-    ToysCounterFiltersVariables,
-    MasterCounterVariables,
-)
-from .queries import (
-    GetMeQuery,
-    GetMasterByUserQuery,
-    GetUserByIDQuery,
-    GetUserByEmailQuery,
-    ToysCatalogQuery,
-    ToysCounterQuery,
-    AllToysCategoriesQuery,
-    AllToysTagsQuery,
-    ToyByIDQuery,
-    MastersCatalogQuery,
-    MastersCounterQuery,
-)
